requirements_dev.txt
pytest
time-machine
pytest-django
coverage
